requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
boto3==1.28.57
Pillow==10.0.1
python-dotenv==1.0.0
google-generativeai==0.1.0
gunicorn==21.2.0 